\docType{package}
\name{clickme-package}
\alias{clickme}
\alias{clickme-package}
\title{Generate JavaScript visualizations directly from R}
\description{
  clickme lets you plot your R data as interactive
  JavaScript visualizations
}
\author{
  Nacho Caballero <\url{http://reasoniamhere.com}>
}
\references{
  Full documentation:
  \url{https://github.com/nachocab/clickme/wiki};
}

