% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/clickme.R, R/clickme_package.R
\docType{package}
\name{clickme}
\alias{clickme}
\alias{clickme-package}
\title{Create a new clickme visualization}
\usage{
clickme(template, ...)
}
\arguments{
\item{template}{template used to generate the visualization}

\item{...}{arguments passed to the template}
}
\description{
\code{clickme()} is used to create a clickme visualization.

Clickme: use R to create interactive visualizations
}
\details{
There are four ways to call \code{clickme}:
\itemize{
   \item \code{clickme(template, arguments)}
   \item \code{clickme(arguments)}
   \item \code{clickme(template)}
   \item \code{clickme()}
  }

The first form is recommended for code that is run from a script. It sets the
template and the arguments to be used.
The second form is recommended for quick command line input. Instead of setting
the template, it uses the previously used template.
The third form sets the template, but doesn't generate any visualization. This is
useful when combined with the second form.
The fourth way returns the current template.
}
\section{Package options}{


Clickme uses the following \code{options} to configure behaviour:

\itemize{

  \item \code{clickme_templates_path}: path where the templates are stored

  \item \code{clickme_output_path}: path where the generated visualizations are saved (along with their dependencies)

}
}
\author{
Nacho Caballero <\url{http://reasoniamhere.com}>
}
\references{
Full documentation: \url{https://github.com/nachocab/clickme/wiki};
}

