\name{set_clickme_path}
\alias{set_clickme_path}
\title{Set the path that contains templates and visualizations}
\usage{
  set_clickme_path(path = getwd())
}
\arguments{
  \item{path}{path that contains the templates folder}
}
\description{
  Set the path that contains templates and visualizations
}

