\name{clickme_vega}
\alias{clickme_vega}
\title{Generates a JavaScript visualization using Vega}
\usage{
  clickme_vega(data, spec, ...)
}
\arguments{
  \item{data}{input data}

  \item{spec}{Name of the vega spec file to use, it must
  match a file within \code{vega/data/spec/}}

  \item{...}{additional arguments for \code{clickme}}
}
\description{
  Generates a JavaScript visualization using Vega
}

