\name{demo_mode}
\alias{demo_mode}
\title{Run a template demo}
\usage{
demo_mode(on = NULL, iframe_src = "src", iframe_height = 800,
  demo_path = "./clickme_demo")
}
\arguments{
  \item{template}{name of template Demo mode makes
  params$dir="./clickme_demo" (in the current directory) it
  calls clickme with cme(...)$iframe()$hide() and it adds a
  random string to the output file}
}
\description{
Run a template demo
}

