% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.R
\name{disjoint_sets}
\alias{disjoint_sets}
\title{Split up two vectors into their intersecting sets}
\usage{
disjoint_sets(a, b, names = c("a", "b", "both"))
}
\arguments{
\item{a}{first vector}

\item{b}{second vector

It returns a list of three elements, those that are only in a, those that are in both, and those that are only in b.}
}
\description{
Split up two vectors into their intersecting sets
}

