% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.R
\name{enumerate}
\alias{enumerate}
\title{Return the elements of a character vector separated by newlines}
\usage{
enumerate(x)
}
\description{
Return the elements of a character vector separated by newlines
}

