% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_functions}
\alias{extract_functions}
\title{Extract function names from a list of placeholder expressions}
\usage{
extract_functions(expressions)
}
\description{
Extract function names from a list of placeholder expressions
}
\keyword{internal}

