\name{generate_visualization}
\alias{generate_visualization}
\title{Generate the JavaScript visualization}
\usage{
  generate_visualization(data, opts)
}
\description{
  Write the input data.frame to a file, for easier
  JavaScript consumption we use capture.output to hide the
  knitr progress bars, how should we deal with errors?
}

