\name{get_opts}
\alias{get_opts}
\title{Get options used by ractive}
\usage{
  get_opts(ractive, data_file_name = NULL,
    html_file_name = NULL)
}
\arguments{
  \item{ractive}{name of the reactive}

  \item{data_file_name}{name of output data file}

  \item{html_file_name}{name of output visualization file}
}
\description{
  Get options used by ractive
}

