\name{get_padding_param}
\alias{get_padding_param}
\title{Get padding around plot}
\usage{
  get_padding_param(opts,
    default = c(top = 10, right = 30, bottom = 30, left = 10))
}
\arguments{
  \item{opts}{ractive options}

  \item{default}{default padding, a vector with top, left,
  bottom and right values}
}
\description{
  Get padding around plot
}

