% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/reload_translators.R
\name{get_sourced_objects}
\alias{get_sourced_objects}
\title{Traverse a list of paths and determine what objects are defined in the
source files that appear in the "translator" folder of each path
It returns a list of sourced objects.
We do this to control what objects get added to the global namespace}
\usage{
get_sourced_objects(paths)
}
\description{
Traverse a list of paths and determine what objects are defined in the
source files that appear in the "translator" folder of each path
It returns a list of sourced objects.
We do this to control what objects get added to the global namespace
}

