% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.R
\name{get_tooltip_format}
\alias{get_tooltip_format}
\title{Return a C-style format given the type of data}
\usage{
get_tooltip_format(variable_value)
}
\description{
Ex: ".2f" for decimal values, "s" for everything else
It also allows to set a custom format
}
\keyword{internal}

