% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.R
\name{get_unique_elements}
\alias{get_unique_elements}
\title{Return the levels of a factor, or the unique elements of a character vector}
\usage{
get_unique_elements(elements)
}
\arguments{
\item{elements}{values}
}
\description{
Return the levels of a factor, or the unique elements of a character vector
}

