% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/xy_to_data.R
\name{get_xy_rownames}
\alias{get_xy_rownames}
\title{Generate appropriate rownames for an object}
\usage{
get_xy_rownames(x)
}
\arguments{
\item{x}{data object}
}
\description{
Generate appropriate rownames for an object
}

