\docType{data}
\name{microarray}
\alias{microarray}
\title{500 genes with the highest fold-change between two conditions}
\format{A data frame with 500 rows and 7 variables}
\description{
  A microarray dataset showing the expression values of 500
  genes The variables include:
}
\details{
  \itemize{ \item logFC. Fold change between two conditions
  (log2) \item magnitude. Average magnitude of the change
  (log2) \item P.Value. t-test-derived p-value \item
  adj.P.Val. Benjamini-Hochberg corrected p-value \item
  probe_name. Agilent v2 probe name \item gene_name. Gene
  symbol \item significance. -log10(adj.P.Val) ... }
}
\keyword{datasets}

