\name{new_ractive}
\alias{new_ractive}
\title{Creates a new ractive}
\usage{
  new_ractive(ractive_name, overwrite = FALSE)
}
\arguments{
  \item{ractive_name}{name of the ractive}

  \item{overwrite}{flag that indicates what to do when
  there is another ractive of the same name, default FALSE}
}
\description{
  This creates the folder structure with the files that a
  ractive needs.
}

