% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/new_template.R
\name{new_template}
\alias{new_template}
\title{Creates a new template}
\usage{
new_template(template_name, coffee = FALSE, replace = FALSE)
}
\arguments{
\item{template_name}{name of the template}

\item{replace}{flag that indicates what to do when there is another template
of the same name, default FALSE}
}
\description{
This creates the folder structure with the files that make up a template.
}

