% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.R
\name{\%notin\%}
\alias{\%notin\%}
\title{Inverse Value Matching}
\usage{
a \%notin\% b
}
\arguments{
\item{a}{a vector}

\item{b}{a vector}
}
\description{
Complement of \code{\%in\%}. Returns the elements of \code{a} that are not in \code{b}.
}

