% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.R
\name{open_html}
\alias{open_html}
\title{Open an HTML file in the browser}
\usage{
open_html(template, ...)
}
\arguments{
\item{template}{template name}

\item{...}{additional fields for \code{get_opts}}
}
\description{
By default it will open \code{get_opts(template)$url}
}

