% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.R
\name{quote_escaped}
\alias{quote_escaped}
\title{Surround with escaped quotes}
\usage{
quote_escaped(data)
}
\arguments{
\item{data}{object to surround with escaped quotes}
}
\description{
Surround with escaped quotes
}

