\name{reload_templates}
\alias{reload_templates}
\title{This function load the R source files in templates/<Template>/R
It first checks that those files are not polluting the global namespace. They are only allowed to define methods to their <Template> object}
\usage{
  reload_templates()
}
\description{
  This function load the R source files in
  templates/<Template>/R It first checks that those files
  are not polluting the global namespace. They are only
  allowed to define methods to their <Template> object
}

