% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/reload_translators.R
\name{reload_translators}
\alias{reload_translators}
\title{Load the translator files in each template}
\usage{
reload_translators()
}
\description{
Translators are only allowed to create one object with the name of their
template and a clickme helper function in the global namespace.
}

