\name{reorder_data_by_color_groups}
\alias{reorder_data_by_color_groups}
\title{Reorder data by color}
\usage{
  reorder_data_by_color_groups(data, params)
}
\arguments{
  \item{data}{data}

  \item{params}{parameters

  It adds a color_groups column and it uses it to reorder the
  data object. This ensures that rows are ordered in a way
  that elements with the first color in the palette are
  rendered on top}
}
\description{
  Reorder data by color
}

