% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.R
\name{server}
\alias{server}
\title{Run a local server}
\usage{
server(path = getOption("clickme_templates_path"), port = 8000)
}
\arguments{
\item{path}{path where server is started}

\item{port}{port used to start the server}
}
\description{
Run a local server
}

