\name{set_root_path}
\alias{set_root_path}
\title{Set the root path}
\usage{
  set_root_path(path)
}
\arguments{
  \item{path}{path to be used as root}
}
\description{
  clickme will use this path as the root for all its
  operations, like generating visualizations. It should
  contain the ractives folder. The server should be run
  from here.
}

