% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.R
\name{test_template}
\alias{test_template}
\title{Test the translator of a given template}
\usage{
test_template(template_name, filter = NULL)
}
\arguments{
\item{template}{name of template}
}
\description{
Test the translator of a given template
}

