\name{translate_data}
\alias{translate_data}
\title{translate is a function defined in template_id/lib/ it might return a JSON object or a file path}
\usage{
  translate_data(data, opts)
}
\description{
  translate is a function defined in template_id/lib/ it
  might return a JSON object or a file path
}

