% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.R
\name{vennize}
\alias{vennize}
\title{Classify the elements of a vector into Venn categories}
\usage{
vennize(a, b, only_in_a = "Only in A", only_in_b = "Only in B",
  in_both = "In both")
}
\description{
Classify the elements of a vector into Venn categories
}

