\docType{package}
\name{clickme}
\alias{clickme}
\alias{clickme-package}
\title{Clickme: use R to create interactive visualizations}
\description{
  Clickme: use R to create interactive visualizations
}
\section{Package options}{
  Clickme uses the following \code{options} to configure
  behaviour:

  \itemize{

  \item \code{clickme_templates_path}: path where the
  templates are stored

  \item \code{clickme_output_path}: path where the
  generated visualizations are saved (along with their
  dependencies)

  }
}
\author{
  Nacho Caballero <\url{http://reasoniamhere.com}>
}
\references{
  Full documentation:
  \url{https://github.com/nachocab/clickme/wiki};
}

