\name{clickme_points}
\alias{clickme_points}
\title{Generates an interactive scatterplot}
\usage{
  clickme_points(x, y = NULL, ...)
}
\arguments{
  \item{x}{x-values}

  \item{y}{y-values}

  \item{...}{additional arguments for \code{clickme}

  \code{x} and \code{y} follow the same behavior as the
  base::plot function. If y is not defined, x is
  interpreted as y. x can be a vector, a list, a
  data.frame, or a matrix.}
}
\description{
  Generates an interactive scatterplot
}

