% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.R
\name{demo_mode}
\alias{demo_mode}
\title{Run a template demo}
\usage{
demo_mode(on = NULL, iframe_src = "src", iframe_height = 800,
  iframe_width = 1000, demo_path = ".")
}
\arguments{
\item{template}{name of template
Demo mode makes params$dir="./clickme_demo" (in the current directory)
it calls clickme with cme(...)$iframe()$hide()
and it adds a random string to the output file}
}
\description{
Run a template demo
}

