\name{get_opts}
\alias{get_opts}
\title{Get options used by ractive}
\usage{
  get_opts(ractive, data_name = NULL,
    html_file_name = NULL, params = NULL)
}
\arguments{
  \item{ractive}{name of the reactive}

  \item{data_name}{name of input data}

  \item{html_file_name}{name of output HTML file}
}
\description{
  Get options used by ractive
}

