\name{get_opts}
\alias{get_opts}
\title{Get options used by the current ractive}
\usage{
  get_opts(ractive, params = NULL, name_mappings = NULL,
    data_prefix = "data", html_file_name = NULL,
    port = 8888)
}
\arguments{
  \item{ractive}{name of the reactive}

  \item{params}{named vector or list containing the
  parameters and values that will be accessible from the
  template}

  \item{name_mappings}{named vector or list of names in the
  input data object and the required names they represent.
  Each element will have the format: name_in_input_data =
  "required_name"}

  \item{data_prefix}{basename of any temporary files
  generated by the ractive, and also used as part of the
  output HTML file name, default "data"}

  \item{html_file_name}{name of the output HTML file that
  contains the visualization, default
  "data_prefix-ractive.html"}

  \item{port}{port used to open a local browser, default
  8888}
}
\description{
  Get options used by the current ractive
}

