\name{get_padding_param}
\alias{get_padding_param}
\title{Get padding around plot}
\usage{
  get_padding_param(opts,
    default = c(top = 100, right = 100, bottom = 100, left = 100))
}
\arguments{
  \item{opts}{template options}

  \item{default}{default padding, a vector with top, left,
  bottom and right values}
}
\description{
  Get padding around plot
}

