\name{get_sourced_objects}
\alias{get_sourced_objects}
\title{Traverses a list of paths and determines what objects are defined in the source files that appear in the "translator" folder of each path
It returns a list of sourced objects.
This is done to make control what objects get added to the global namespace}
\usage{
  get_sourced_objects(paths)
}
\description{
  Traverses a list of paths and determines what objects are
  defined in the source files that appear in the
  "translator" folder of each path It returns a list of
  sourced objects. This is done to make control what
  objects get added to the global namespace
}

