\name{new_template}
\alias{new_template}
\title{Creates a new template}
\usage{
  new_template(template_name, overwrite = FALSE)
}
\arguments{
  \item{template_name}{name of the template}

  \item{overwrite}{flag that indicates what to do when
  there is another template of the same name, default
  FALSE}
}
\description{
  This creates the folder structure with the files that a
  template needs.
}

