\name{reorder_data_by_colorize}
\alias{reorder_data_by_colorize}
\title{Reorder data by color}
\usage{
  reorder_data_by_colorize(data, params)
}
\arguments{
  \item{data}{data}

  \item{params}{parameters

  It adds a colorize column and it uses it to reorder the
  data object. This ensures that rows are ordered in a way
  that elements with the first color in the palette are
  rendered on top}
}
\description{
  Reorder data by color
}

