\name{scale_type}
\alias{scale_type}
\title{Type of scale}
\usage{
  scale_type(elements)
}
\arguments{
  \item{elements}{values

  If elements is numeric and has a length greater than one,
  it returns "quantitative". If elements is NULL, or not
  numeric, or has a length of one, it returns
  "categorical".}
}
\description{
  Type of scale
}

